Lift the login/session endpoints out of web_server
api/login.py takes /api/auth/login, logout, recovery-question and
recovery-reset, and the login brute-force limiter moves with them -
web_server imports the limiter back because api/user_profiles gets it
injected (admin endpoints clear lockouts).

named api/login.py, not api/auth.py: that name is TAKEN by the public
REST API's key authentication, and the first attempt overwrote it. the
routemap diff caught it - every /api/v1 route vanished - which is
exactly the failure mode that check exists for. restored from the
index, re-carved under the free name, routemap identical again (v1
included). login endpoint + gate suites pass (19).
