+++ web_server.py
@@ -553,7 +553,9 @@ def inject_webui_assets():
 # PINs from one IP trips it — correct entry clears it instantly).
 from core.security.rate_limit import AttemptLimiter as _AttemptLimiter
 _launch_pin_limiter = _AttemptLimiter(max_attempts=10, window_seconds=300)
-_login_limiter = _AttemptLimiter(max_attempts=10, window_seconds=300)
+# the login/recovery limiter moved to api/login.py with its routes; imported
+# back because api/user_profiles gets it injected (admin clears lockouts)
+from api.login import login_limiter as _login_limiter
 
 
 def _require_login_enabled():
@@ -27626,109 +27628,6 @@ def hydrate_beatport_bubbles():
 # --- Profile API Endpoints ---
 
 # /api/profiles* endpoints: lifted to api/user_profiles.py.
-@app.route('/api/auth/login', methods=['POST'])
-def auth_login():
-    """Username/password login (opt-in login mode). Username = profile name.
-    Brute-force limited per IP; a profile with no password set can't log in."""
-    try:
-        _ip = request.remote_addr or 'unknown'
-        _now = time.time()
-        _locked, _retry_after = _login_limiter.is_locked(_ip, _now)
-        if _locked:
-            return (jsonify({'success': False, 'error': 'Too many attempts — please wait and try again'}),
-                    429, {'Retry-After': str(_retry_after)})
-
-        data = request.json or {}
-        username = (data.get('username') or '').strip()
-        password = data.get('password') or ''
-        if not username or not password:
-            return jsonify({'success': False, 'error': 'Username and password required'}), 400
-
-        database = get_database()
-        profile = database.get_profile_by_name(username)
-        # Same generic error + a recorded failure whether the name or password is
-        # wrong — don't leak which names exist.
-        if not profile or not database.verify_profile_password(profile['id'], password):
-            _login_limiter.record_failure(_ip, _now)
-            return jsonify({'success': False, 'error': 'Invalid username or password'}), 401
-
-        _login_limiter.record_success(_ip)
-        session['login_authenticated'] = True
-        session['profile_id'] = profile['id']
-        # A fresh login also clears any stale launch-PIN flag.
-        session.pop('launch_pin_verified', None)
-        return jsonify({'success': True, 'profile': {
-            'id': profile['id'], 'name': profile['name'], 'is_admin': profile['is_admin'],
-        }})
-    except Exception as e:
-        return jsonify({'success': False, 'error': str(e)}), 500
-
-
-@app.route('/api/auth/logout', methods=['POST'])
-def auth_logout():
-    """Log out — clears the authenticated session."""
-    try:
-        session.pop('login_authenticated', None)
-        session.pop('profile_id', None)
-        session.pop('launch_pin_verified', None)
-        return jsonify({'success': True})
-    except Exception as e:
-        return jsonify({'success': False, 'error': str(e)}), 500
-
-
-@app.route('/api/auth/recovery-question', methods=['GET'])
-def auth_recovery_question():
-    """Return the recovery security-question for a username (forgot-password flow).
-    Generic when the user/question is absent — don't confirm which names exist."""
-    try:
-        username = (request.args.get('username') or '').strip()
-        database = get_database()
-        profile = database.get_profile_by_name(username) if username else None
-        question = database.get_profile_recovery_question(profile['id']) if profile else None
-        if not question:
-            return jsonify({'success': False, 'error': 'No recovery question available'}), 404
-        return jsonify({'success': True, 'question': question})
-    except Exception as e:
-        return jsonify({'success': False, 'error': str(e)}), 500
-
-
-@app.route('/api/auth/recovery-reset', methods=['POST'])
-def auth_recovery_reset():
-    """Reset a login password by answering the recovery question. Brute-force
-    limited; a correct answer sets the new password and authenticates the session."""
-    try:
-        _ip = request.remote_addr or 'unknown'
-        _now = time.time()
-        _locked, _retry_after = _login_limiter.is_locked(_ip, _now)
-        if _locked:
-            return (jsonify({'success': False, 'error': 'Too many attempts — please wait and try again'}),
-                    429, {'Retry-After': str(_retry_after)})
-
-        data = request.json or {}
-        username = (data.get('username') or '').strip()
-        answer = data.get('answer') or ''
-        new_password = data.get('new_password') or ''
-        if not username or not answer or not new_password:
-            return jsonify({'success': False, 'error': 'Username, answer and new password are required'}), 400
-        if len(new_password) < 6:
-            return jsonify({'success': False, 'error': 'New password must be at least 6 characters'}), 400
-
-        database = get_database()
-        profile = database.get_profile_by_name(username)
-        if not profile or not database.verify_profile_recovery_answer(profile['id'], answer):
-            _login_limiter.record_failure(_ip, _now)
-            return jsonify({'success': False, 'error': 'Incorrect answer'}), 401
-
-        _login_limiter.record_success(_ip)
-        database.set_profile_password(profile['id'], new_password)
-        session['login_authenticated'] = True
-        session['profile_id'] = profile['id']
-        session.pop('launch_pin_verified', None)
-        return jsonify({'success': True})
-    except Exception as e:
-        return jsonify({'success': False, 'error': str(e)}), 500
-
-
 # Per-profile service-connection helpers: moved to api/user_profiles.py with
 # the profile routes that use them.
 from api.user_profiles import _get_lb_credentials_for_profile  # one outside caller remains
@@ -35670,6 +35569,10 @@ _cfg_verif(get_database_=get_database, config_manager_=config_manager,
            download_orchestrator_getter=lambda: download_orchestrator,
            matching_engine_getter=lambda: matching_engine)
 app.register_blueprint(_bp_verif())
+# Login/session endpoints (api/login.py).
+from api.login import configure as _cfg_login, create_blueprint as _bp_login
+_cfg_login(get_database_=get_database)
+app.register_blueprint(_bp_login())
 # Quarantine review endpoints (api/quarantine.py).
 from api.quarantine import configure as _cfg_quar, create_blueprint as _bp_quar
 _cfg_quar(config_manager_=config_manager, docker_resolve_path_=docker_resolve_path,
